// src/app/types/Employee/Employee-createDTO.ts
export type EmployeeCreateDTO = {
  name: string;
  last_name: string;
  middle_name?: string | null;
  snils?: string | null;
  birth_date?: string | null;
  organization_id: number;
  grade?: number | null;
  phone?: string | null;
  email?: string | null;
  education: string;
  is_active?: boolean;
}